No longer ignore settings.cfg
While it is possible that someone may inadvertently store settings
in this file that should not be committed to a repository, but
currently it does not contain any.

If this file is not committed, then clones based on this cookiecutter
will fail due to not being able to find a settings.cfg.